feat: separate tokenizing from crawling
